server/model-server/requirements.txt
flask
flask-cors
tensorflow
python-dotenv
Pillow
numpy
requests
